Connector method added to connect with localdb

+++ More Exercises/JurassicPark/JurassicPark/People/Employees/Employee.cs
@@ -2,6 +2,6 @@
 
 internal class Employee : Person
 {
-    internal int Salary { get; set; }
+    internal decimal Salary { get; set; }
     internal string Department { get; set; }
 }
